refactor: add unit-ban module
