+++ core-commonui/unit-ban/src/commonMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/unit/ban/di/Utils.kt
@@ -1,0 +1,11 @@
+package com.github.diegoberaldin.raccoonforlemmy.unit.ban.di
+
+import com.github.diegoberaldin.raccoonforlemmy.unit.ban.BanUserMviModel
+
+expect fun getBanUserViewModel(
+    userId: Int,
+    communityId: Int,
+    newValue: Boolean = true,
+    postId: Int? = null,
+    commentId: Int? = null,
+): BanUserMviModel
