+++ settings.gradle.kts
@@ -42,3 +42,4 @@ include(":feature-inbox")
 include(":feature-search")
 include(":feature-profile")
 include(":feature-settings")
+include(":core-commonui:unit-ban")
